refactor: add StatusCode.cpp
