+++ src/ProcessOps.cpp
@@ -1,5 +1,4 @@
 #include "ProcessOps.hpp"
-#include <cstdio>
 
 /* ====== CONSTRUCTOR/DESTRUCTOR ====== */
